WorkerId is now u32

+++ crates/tako/src/lib.rs
@@ -8,7 +8,7 @@ pub mod server;
 pub mod transfer;
 pub mod worker;
 
-pub type WorkerId = u64; // Maybe non-zero type for optimizing Option<WorkerId>?
+pub type WorkerId = u32; // Maybe non-zero type for optimizing Option<WorkerId>?
 pub type TaskId = u64;
 pub type TaskTypeId = u32;
 pub type OutputId = u32;
